requirements.txt

requests>=2.31.0
Pillow
fastapi
uvicorn[standard]
httpx>=0.27.0
